Workaround: OpenSSL session ticket client bug
The OpenSSL client does not call the new session callback when
sessions are obsoleted by the sever issuing a new session ticket
(possible after a full handshake!).  The client may keep doing full
handshakes until it flushes the session from its own cache.

We detect this and make the call on OpenSSL's behalf.  This is
future-proofed against OpenSSL getting it right by detecting that
too and avoiding duplication.